modify alsa initialization
This deals with the alsa devices having different names. The function will iterate over all active simple elements and select the first one it finds that can adjust the playback volume. If it can't find any, it prints an error message and cleans up. All other audio is disabled, so it should always detect PWM audio.